+++ src/refactoring/3_widgets/layouts/index.ts
@@ -1,2 +1,1 @@
-// 레이아웃 컴포넌트들을 내보냅니다
-export { RootLayout } from "./root-layout";
+export * from "./ui/root-layout";
