+++ src/refactoring/3_widgets/layouts/ui/root-layout.tsx
@@ -1,13 +1,8 @@
-import { Outlet, useLocation, useNavigate } from "react-router-dom";
+import { Outlet } from "react-router-dom";
+import { useAdminCartToggle } from "../hooks/use-admin-cart-toggle";
 
 export const RootLayout = () => {
-  const location = useLocation();
-  const navigate = useNavigate();
-
-  const isAdminPage = location.pathname === "/admin";
-
-  const togglePath = isAdminPage ? "/cart" : "/admin";
-  const toggleText = isAdminPage ? "장바구니 페이지로" : "관리자 페이지로";
+  const { toggleText, handleToggleClick } = useAdminCartToggle();
 
   return (
     <div className="min-h-screen bg-gray-100">
@@ -15,7 +10,7 @@ export const RootLayout = () => {
         <div className="container mx-auto flex justify-between items-center">
           <h1 className="text-2xl font-bold">쇼핑몰 관리 시스템</h1>
           <button
-            onClick={() => navigate(togglePath)}
+            onClick={handleToggleClick}
             className="bg-white text-blue-600 px-4 py-2 rounded hover:bg-blue-100"
           >
             {toggleText}
